Added http error logging
