+++ src/app/modules/project/overview/overview.component.ts
@@ -67,6 +67,7 @@ export class OverviewComponent implements OnInit {
           if (error.status !== 404) {
             console.log("Could not retrieve the projects");
           }
+          throw error;
         }
       );
   }
